Important discoveries about x86_64 syscall, signals, process syncing
First, I finally figured out that "normal" Linux x86_64 syscalls don't
save an entire stack frame, and thus do_fork() was failing because IP
and SP we not the values of the calling process. The solution was to
have dput/dget be invoked like sys_fork where a full stack frame is
indeed saved. See entry_64.S PTREGSCALL.

I realized I *still* wasn't using the waitqueue properly to wait for a
process to synchronize...I now do a for loop until the condition is met
to continue execution.

As a part of the above, I discovered that killing (SIGINT, SIGKILL) a
process waiting in a waitqueue loop would never exit the loop.
Apparently, do_signal(), which handles a SIGKILL, etc, is only run when
returning into user mode, not back into a syscall. I don't yet have a
fix, but the BUG is documented in determinism.c.